refactor (product, member) : member, product dependency partition
