+++ src/main/java/com/levelUpToast/levelUpToast/function/member/memberCheck/MemberCheck.java
@@ -1,0 +1,10 @@
+package com.levelUpToast.levelUpToast.function.member.memberCheck;
+
+import com.levelUpToast.levelUpToast.config.exception.LevelUpToastEx;
+import com.levelUpToast.levelUpToast.domain.model.member.Member;
+
+public interface MemberCheck {
+    Boolean idPresentCheck(String id);
+
+    Member checkMember(String requestSeq) throws LevelUpToastEx;
+}
