+++ src/main/java/com/levelUpToast/levelUpToast/function/member/memberCreate/MemberCreate.java
@@ -1,0 +1,4 @@
+package com.levelUpToast.levelUpToast.function.member.memberCreate;
+
+public interface MemberCreate {
+}
