+++ src/main/java/com/levelUpToast/levelUpToast/function/product/productCheck/ProblemCheck.java
@@ -1,4 +1,4 @@
-package com.levelUpToast.levelUpToast.function.productInspection;
+package com.levelUpToast.levelUpToast.function.product.productCheck;
 
 import com.levelUpToast.levelUpToast.config.exception.LevelUpToastEx;
 import com.levelUpToast.levelUpToast.domain.model.member.Member;
@@ -7,7 +7,6 @@ import com.levelUpToast.levelUpToast.domain.model.product.Product;
 import java.util.Optional;
 
 public interface ProblemCheck {
-    Member checkMember(String requestSeq) throws LevelUpToastEx;
 
     void checkProductSEQ(Long originalProductVendor, Long requestSeq) throws LevelUpToastEx;
