+++ src/main/java/com/levelUpToast/levelUpToast/service/login/loginServiceImpl/SimpleLoginService.java
@@ -21,7 +21,7 @@ public class SimpleLoginService implements LoginService {
 
     @Override
     public Member login(String loginId, String password) throws LevelUpToastEx {
-        Optional<Member> findMem = memberRepository.findByloginId(loginId);
+        Optional<Member> findMem = memberRepository.findByLoginId(loginId);
         if (findMem.isEmpty()) {
             log.warn("[LoginService log] : 미존재 ID 로그인 요청  ID = {}, PW = {}", loginId, password);
             throw new LevelUpToastEx("동일한 ID가 존재하지 않습니다.", 2);
